[S06P22A407-123] feat: 피드백 답변 작성

+++ backend/src/main/java/com/daram/dotore/api/request/AnswerReq.java
@@ -1,7 +1,6 @@
 package com.daram.dotore.api.request;
 
 import io.swagger.annotations.ApiModel;
-import java.math.BigInteger;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -10,7 +9,7 @@ import lombok.Setter;
 @ApiModel("AnswerRequest")
 public class AnswerReq {
 
-    private BigInteger articleno;
+    private int articleno;
     private String writer;
     private String description;
     private String img_url;
